Lot of code fixing and modernization

+++ Network/mdz_net_chains/socket_chain.cpp
@@ -179,21 +179,21 @@ int Socket_Chain::shutdownSocket(int mode)
     return curSocket->shutdownSocket(mode);
 }
 
-int Socket_Chain::partialRead(void *data, const uint32_t & datalen)
+ssize_t Socket_Chain::partialRead(void *data, const uint32_t & datalen)
 {
     if (socketLayers.size() == 0 && !baseSocket) return -1;
     Mantids::Network::Sockets::Socket_StreamBase * curSocket = !socketLayers.size()? baseSocket : ((sChainVectorItem *)socketLayers[socketLayers.size()-1])->sock[0];
-    int x = curSocket->partialRead(data,datalen);
+    ssize_t x = curSocket->partialRead(data,datalen);
     if (x<=0)
         return x;
     return x;
 }
 
-int Socket_Chain::partialWrite(const void *data, const uint32_t &datalen)
+ssize_t Socket_Chain::partialWrite(const void *data, const uint32_t &datalen)
 {
     if (socketLayers.size() == 0 && !baseSocket) return -1;
     Mantids::Network::Sockets::Socket_StreamBase * curSocket = !socketLayers.size()? baseSocket : ((sChainVectorItem *)socketLayers[socketLayers.size()-1])->sock[0];
-    int x = curSocket->partialWrite(data,datalen);
+    ssize_t x = curSocket->partialWrite(data,datalen);
     if (x<=0)
         return x;
     return x;
